+++ src/components/dialog/dialog.component.tsx
@@ -4,7 +4,6 @@ import createGuid from "../../__internal__/utils/helpers/guid";
 import Modal, { ModalProps } from "../modal";
 import Heading from "../heading";
 import tagComponent, { TagProps } from "../../__internal__/utils/helpers/tags";
-import Logger from "../../__internal__/utils/logger";
 
 import {
   StyledDialog,
@@ -29,10 +28,12 @@ export interface ContentPaddingInterface {
   px?: PaddingValues;
 }
 
-let deprecatedClassNameWarningShown = false;
-
 export interface DialogProps extends ModalProps, TagProps {
-  /** Custom class name  */
+  /**
+   * @private
+   * @ignore
+   * @internal
+   * Sets className for component. INTERNAL USE ONLY. */
   className?: string;
   /** Prop to specify the aria-describedby property of the Dialog component */
   "aria-describedby"?: string;
@@ -131,13 +132,6 @@ export const Dialog = forwardRef<DialogHandle, DialogProps>(
     },
     ref,
   ) => {
-    if (!deprecatedClassNameWarningShown && className) {
-      Logger.deprecate(
-        "The 'className' prop has been deprecated and will soon be removed from the 'Dialog' component.",
-      );
-      deprecatedClassNameWarningShown = true;
-    }
-
     const locale = useLocale();
 
     const containerRef = useRef<HTMLDivElement>(null);