+++ src/components/icon/icon.style.ts
@@ -1,26 +1,24 @@
 import styled, { css } from "styled-components";
+
 import { shade } from "polished";
+
 import { margin } from "styled-system";
 
-import iconUnicodes from "./icon-unicodes";
 import baseTheme, { ThemeObject } from "../../style/themes/base";
-import iconConfig from "./icon-config";
+import addFocusStyling from "../../style/utils/add-focus-styling";
+import styledColor from "../../style/utils/color";
+import getColorValue from "../../style/utils/get-color-value";
 import browserTypeCheck, {
   isSafari,
 } from "../../__internal__/utils/helpers/browser-type-check";
-import styledColor from "../../style/utils/color";
-import getColorValue from "../../style/utils/get-color-value";
+
+import iconConfig from "./icon-config";
 import { IconType } from "./icon-type";
-import addFocusStyling from "../../style/utils/add-focus-styling";
+import iconUnicodes from "./icon-unicodes";
 
 export type BackgroundShape = "circle" | "rounded-rect" | "square";
 
-export type BgSize =
-  | "extra-small"
-  | "small"
-  | "medium"
-  | "large"
-  | "extra-large";
+export type BgSize = "small" | "medium" | "large" | "extra-large";
 
 export type FontSize = "small" | "medium" | "large" | "extra-large";
 
@@ -59,11 +57,10 @@ export interface StyledIconInternalProps {
 
 function adjustIconBgSize(fontSize?: FontSize, bgSize?: BgSize) {
   const sizeValues: Record<BgSize | FontSize, number> = {
-    "extra-small": 1,
-    small: 2,
-    medium: 3,
-    large: 4,
-    "extra-large": 5,
+    small: 1,
+    medium: 2,
+    large: 3,
+    "extra-large": 4,
   };
 
   if (fontSize && bgSize) {
@@ -81,7 +78,7 @@ function adjustIconBgSize(fontSize?: FontSize, bgSize?: BgSize) {
     return iconConfig.backgroundSize[bgSize];
   }
 
-  /* The below is ignored as removing it may cause regressions as some components import StyledIcon directly from this file 
+  /* The below is ignored as removing it may cause regressions as some components import StyledIcon directly from this file
   however it cannot be tested in the Icon tests as these props always have a value. */
   /* istanbul ignore next */
   return bgSize ? iconConfig.backgroundSize[bgSize] : undefined;
@@ -180,7 +177,7 @@ const StyledIcon = styled.span<StyledIconProps & StyledIconInternalProps>`
         css`
           margin-top: -6px;
         `}
-        
+
         display: block;
       }